CommonJS module importing from a package with "type": "module"
In importing/main.js
```
const { int } = require("exporting");
                        ^^^^^^^^^^^
```
Gets error:

    The current file is a CommonJS module whose imports will produce 'require' calls; however, the referenced file is an ECMAScript module and cannot be imported with 'require'. Consider writing a dynamic 'import("exporting")' call instead.
  To convert this file to an ECMAScript module, change its file extension to '.mjs' or create a local package.json file with `{ "type": "module" }`.ts(1479)

The type of `int` is correctly resolved.

This seems to have been caused by the exporting package contains in its package.json
```
    "type": "module"
```

The error is not triggered when `"type"` is `"commonjs"` (or omitted).

Perhaps the error logic should also consult the extension of the actually imported module, i.e. `"main.cjs"`. The `"type"` field should only be consulted when the exported module has a .js extension. In this case `"exports"."."."require"` references a cjs file. It is obviously not an ES module.